refactor: removes static file approach
