+++ src/lib/remark-plugins/rewrite-tutorial-links/utils/index.ts
@@ -1,2 +1,2 @@
-export { generateTutorialMap } from './generate-tutorial-map'
+export { getTutorialMap } from './get-tutorial-map'
 export { handleCollectionLink, handleTutorialLink } from './handle-learn-links'
